Create batchMykrobe_genotype.sh
#MacOS shell script for Mykrobe Predictor
This script allows to run Mykrobe genotype (see https://github.com/iqbal-lab/Mykrobe-predictor) in parallel for a batch of sequence read files from the macOS bash shell. 

## Dependencies:
The script is dependent on the command line version of Mykrobe predictor (see https://github.com/iqbal-lab/Mykrobe-predictor) with its dependencies. For the -s (summarise) option, the python script json_summary.py from this repository is required (requires python 2.7 or higher). 
Furthermore, the path to the mccortex (line 11, see Mykrobe documentation) and the path to the summary file in this repository (line 154) needs to be adjusted to your system. 

## Input (required): 
-i Path to a folder containing fastq.gz files
-o path to a folder containing the .json files with the program output
-p the panel of reference sequences to test for

## Input (optional): 
-f name of output file for summary (.tsv)

## Comments on the script
The script  can probably easily be adapted for linux, but I did not have time to do that so far. Also, there might be much more elegant code lines to achieve the same - my apologies, I am not a programmer.